+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Features:
 - Gauge can set request limit to avoid total number of pending requests overload the chain network.
 - Added alternative passive request strategy "Full Passive Demand Strategy" (experimental, need to enable in configuration).
 - Added alternative compact font for input/output count rendering in recipe settings screen (configurable)
+- JEI/EMI support for drag-to-set empty gauge item while in Controller Screen
 - Create Production Task icon in Stock Keeper GUI will now include associated gauge ingredients and target address in tooltip.
 
 Changes: